+++ app/core/src/main/scala/com/lollypop/runtime/instructions/infrastructure/DeclareTableFrom.scala
@@ -4,41 +4,29 @@ import com.lollypop.language.models.{Atom, Queryable, TableModel}
 import com.lollypop.runtime._
 import com.lollypop.runtime.datatypes.TableType.TableTypeRefExtensions
 import com.lollypop.runtime.devices.RowCollectionZoo.createTempTable
-import com.lollypop.runtime.devices.{Field, FieldMetadata}
-import com.lollypop.runtime.instructions.queryables.RowsOfValues
 import lollypop.io.IOCost
 
 /**
  * declare table ... from statement
- * @param ref         the given [[Atom database object reference]]
- * @param tableModel  the given [[TableModel table]]
- * @param from        the given [[Queryable queryable]]
+ * @param ref         the provided [[Atom database object reference]]
+ * @param tableModel  the provided [[TableModel table]]
+ * @param from        the source [[Queryable queryable]]
  * @param ifNotExists if true, the operation will not fail when the entity exists
  * @author lawrence.daniels@gmail.com
  */
 case class DeclareTableFrom(ref: Atom, tableModel: TableModel, from: Queryable, ifNotExists: Boolean)
-  extends TableCreation {
+  extends TableCreationFrom {
+  protected def actionVerb: String = "declare"
 
   override def execute()(implicit scope: Scope): (Scope, IOCost, IOCost) = {
     // attempt to declare the table
     val _type = tableModel.toTableType
-    val out = createTempTable(_type.columns)
+    val device = createTempTable(_type.columns)
 
     // insert the rows
-    val cost1 = from match {
-      case RowsOfValues(values) =>
-        val fields = tableModel.columns.map(c => Field(name = c.name, metadata = FieldMetadata(), value = c.defaultValue))
-        out.insertRows(fields.map(_.name), values)
-      case queryable =>
-        val (_, cost2, device1) = queryable.search(scope)
-        cost2 ++ out.insert(device1)
-    }
+    val cost1 = insertRows(device)
 
-    (scope.withVariable(Variable(name = ref.name, _type, initialValue = out)), cost1, cost1)
+    (scope.withVariable(Variable(name = ref.name, _type, initialValue = device)), cost1, cost1)
   }
-
-  override def toSQL: String = s"${super.toSQL} containing ${from.toSQL}"
-
-  protected def actionVerb: String = "declare"
 
 }
